+++ .eslintrc.js
@@ -13,7 +13,7 @@ module.exports = {
       files: ["*.ts"],
       parser: "@typescript-eslint/parser",
       parserOptions: {
-        project: ["tsconfig.json"],
+        project: ["tsconfig.json", "./packages/**/tsconfig.json"],
       },
       plugins: [
         "eslint-plugin-import",