+++ packages/cli/src/lib/helpers/manifest.ts
@@ -16,7 +16,6 @@ import {
 } from "@web3api/core-js";
 import { writeFileSync, normalizePath } from "@web3api/os-js";
 import { Schema as JsonSchema } from "jsonschema";
-import YAML from "js-yaml";
 import path from "path";
 import fs from "fs";
 
@@ -273,7 +272,7 @@ export async function outputManifest(
     };
 
     const sanitizedManifest = removeUndefinedProps(manifest);
-    const str = YAML.safeDump(sanitizedManifest);
+    const str = JSON.stringify(sanitizedManifest);
 
     if (!str) {
       const noDumpMessage = intlMsg.lib_helpers_manifest_unableToDump({