+++ packages/js/plugins/ethereum/src/__tests__/integration/src/query/index.ts
@@ -31,7 +31,7 @@ export function callContractView(
     method: input.method,
     args: input.args,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function callContractStatic(
@@ -43,7 +43,7 @@ export function callContractStatic(
     args: input.args,
     connection: input.connection,
     txOverrides: input.txOverrides
-  });
+  }).unwrap();
 }
 
 export function encodeParams(
@@ -52,7 +52,7 @@ export function encodeParams(
   return Ethereum_Query.encodeParams({
     types: input.types,
     values: input.values
-  });
+  }).unwrap();
 }
 
 export function encodeFunction(
@@ -61,7 +61,7 @@ export function encodeFunction(
   return Ethereum_Query.encodeFunction({
     method: input.method,
     args: input.args
-  });
+  }).unwrap();
 }
 
 export function getSignerAddress(
@@ -69,7 +69,7 @@ export function getSignerAddress(
 ): string {
   return Ethereum_Query.getSignerAddress({
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function getSignerBalance(
@@ -78,7 +78,7 @@ export function getSignerBalance(
   return Ethereum_Query.getSignerBalance({
     blockTag: input.blockTag,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function getSignerTransactionCount(
@@ -87,7 +87,7 @@ export function getSignerTransactionCount(
   return Ethereum_Query.getSignerTransactionCount({
     blockTag: input.blockTag,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function getGasPrice(
@@ -95,7 +95,7 @@ export function getGasPrice(
 ): BigInt {
   return Ethereum_Query.getGasPrice({
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function estimateTransactionGas(
@@ -104,7 +104,7 @@ export function estimateTransactionGas(
   return Ethereum_Query.estimateTransactionGas({
     tx: input.tx,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function estimateContractCallGas(
@@ -116,7 +116,7 @@ export function estimateContractCallGas(
     args: input.args,
     connection: input.connection,
     txOverrides: input.txOverrides
-  });
+  }).unwrap();
 }
 
 export function checkAddress(
@@ -124,7 +124,7 @@ export function checkAddress(
 ): bool {
   return Ethereum_Query.checkAddress({
     address: input.address
-  });
+  }).unwrap();
 }
 
 export function toWei(
@@ -132,7 +132,7 @@ export function toWei(
 ): BigInt {
   return Ethereum_Query.toWei({
     eth: input.eth
-  });
+  }).unwrap();
 }
 
 export function toEth(
@@ -140,7 +140,7 @@ export function toEth(
 ): String {
   return Ethereum_Query.toEth({
     wei: input.wei
-  });
+  }).unwrap();
 }
 
 export function awaitTransaction(
@@ -151,7 +151,7 @@ export function awaitTransaction(
     confirmations: input.confirmations,
     timeout: input.timeout,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function waitForEvent(
@@ -163,7 +163,7 @@ export function waitForEvent(
     args: input.args,
     timeout: input.timeout,
     connection: input.connection
-  });
+  }).unwrap();
 }
 
 export function getNetwork(
@@ -171,5 +171,5 @@ export function getNetwork(
 ): Ethereum_Network {
   return Ethereum_Query.getNetwork({
     connection: input.connection
-  });
+  }).unwrap();
 }