+++ packages/test-cases/cases/bind/sanity/output/wasm-as/mutation/imported/TestImport_Mutation/index.ts
@@ -2,7 +2,8 @@ import {
   w3_subinvoke,
   Nullable,
   BigInt,
-  JSON
+  JSON,
+  Result
 } from "@web3api/wasm-as";
 import {
   serializeimportedMethodArgs,
@@ -18,7 +19,9 @@ export class TestImport_Mutation {
 
   public static uri: string = "testimport.uri.eth";
 
-  public static importedMethod(input: Input_importedMethod): Types.TestImport_Object | null {
+  public static importedMethod(
+    input: Input_importedMethod
+  ): Result<Types.TestImport_Object | null, string> {
     const args = serializeimportedMethodArgs(input);
     const result = w3_subinvoke(
       "testimport.uri.eth",
@@ -26,10 +29,21 @@ export class TestImport_Mutation {
       "importedMethod",
       args
     );
-    return deserializeimportedMethodResult(result);
+
+    if (result.isErr) {
+      return Result.Err<Types.TestImport_Object | null, string>(
+        result.unwrapErr()
+      );
+    }
+
+    return Result.Ok<Types.TestImport_Object | null, string>(
+      deserializeimportedMethodResult(result.unwrap())
+    );
   }
 
-  public static anotherMethod(input: Input_anotherMethod): i32 {
+  public static anotherMethod(
+    input: Input_anotherMethod
+  ): Result<i32, string> {
     const args = serializeanotherMethodArgs(input);
     const result = w3_subinvoke(
       "testimport.uri.eth",
@@ -37,6 +51,15 @@ export class TestImport_Mutation {
       "anotherMethod",
       args
     );
-    return deserializeanotherMethodResult(result);
+
+    if (result.isErr) {
+      return Result.Err<i32, string>(
+        result.unwrapErr()
+      );
+    }
+
+    return Result.Ok<i32, string>(
+      deserializeanotherMethodResult(result.unwrap())
+    );
   }
 }