+++ packages/wasm/as/assembly/msgpack/WriteEncoder.ts
@@ -2,9 +2,9 @@ import { DataView } from "./DataView";
 import { Format } from "./Format";
 import { Nullable } from "./Nullable";
 import { Write } from "./Write";
-import { BigInt } from "../BigInt";
-import { Context } from "./Context";
-import { JSON } from "../JSON";
+import { BigInt } from "../math";
+import { Context } from "../debug";
+import { JSON } from "../json";
 
 export class WriteEncoder extends Write {
   private readonly _context: Context;